requirements.txt
alembic==1.7.6
async-timeout==4.0.2
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Deprecated==1.2.13
dnspython==2.2.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-UUID==0.2
Flask-WTF==1.0.0
gevent==21.12.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.20.0
redis==4.2.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
uritemplate==4.1.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
wrapt==1.14.0
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.4.0
